rataura2/rataura2/requirements.txt
# Livekit dependencies
livekit-agents[google,elevenlabs,openai,silero,deepgram,cartesia,turn-detector,rag,noise-cancellation]~=1.0rc
livekit-api

# Database
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Graph management
networkx>=3.0
matplotlib>=3.5.0

# Utilities
python-dotenv>=0.19.0
loguru>=0.5.3

